+++ src/main/java/pl/car_dealership/infrastructure/database/repository/jpa/ServiceMechanicJpaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import pl.car_dealership.infrastructure.database.entity.ServiceMechanicEntity;
 
 @Repository
-public interface ServiceMechanicJpaRepository extends JpaRepository<ServiceMechanicEntity, Integer> {
+public interface ServiceMechanicJpaRepository extends JpaRepository<ServiceMechanicEntity, Long> {
 
 
 }
