+++ src/main/java/pl/car_dealership/infrastructure/database/repository/jpa/ServicePartJpaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import pl.car_dealership.infrastructure.database.entity.ServicePartEntity;
 
 @Repository
-public interface ServicePartJpaRepository extends JpaRepository<ServicePartEntity, Integer> {
+public interface ServicePartJpaRepository extends JpaRepository<ServicePartEntity, Long> {
 
 
 }
